refactor: rename commandinfrastructure package
