+++ internal/command/infrastructure/mapper.go
@@ -1,4 +1,4 @@
-package commandinfrastructure
+package infrastructure
 
 import (
 	"gomander/internal/command/domain"
